requirements.txt
-i https://pypi.org/simple
alembic==1.11.1; python_version >= '3.7'
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.6.2; python_version >= '3.7'
click==8.1.6; python_version >= '3.7'
dnspython==2.4.1; python_version >= '3.8' and python_version < '4.0'
email-validator==2.0.0.post2
flask==2.3.2; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-migrate==4.0.4
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
pymysql==1.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.19; python_version >= '3.7'
typing-extensions==4.7.1; python_version >= '3.7'
werkzeug==2.3.6; python_version >= '3.8'
wtforms==3.0.1; python_version >= '3.7'
